BusinessTripDAO 인터페이스 추가, BusinessTripMapper NAMESPACE 변경
