+++ src/main/java/com/woosan/hr_system/schedule/service/BusinessTripServiceImpl.java
@@ -1,19 +1,11 @@
 package com.woosan.hr_system.schedule.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.woosan.hr_system.auth.model.UserSessionInfo;
 import com.woosan.hr_system.schedule.dao.BusinessTripDAO;
 import com.woosan.hr_system.schedule.model.BusinessTrip;
-import com.woosan.hr_system.schedule.model.Schedule;
 import lombok.extern.slf4j.Slf4j;
-import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import javax.naming.Context;
 import javax.naming.NamingException;
@@ -21,7 +13,10 @@ import javax.naming.directory.Attributes;
 import javax.naming.directory.DirContext;
 import javax.naming.directory.InitialDirContext;
 import java.time.LocalDateTime;
-import java.util.*;
+import java.util.Hashtable;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @Slf4j
 @Service
